requirements.txt
certifi==2021.10.8
flatbuffers==2.0
numpy==1.21.5
onnxruntime==1.10.0
opencv-python==4.5.4.60
Pillow==8.4.0
protobuf==3.19.1
pyclipper==1.3.0.post2
Shapely==1.8.0
six==1.16.0
